+++ custom_components/access_control/button.py
@@ -8,8 +8,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import DOMAIN
-from .entity import AccessEntity
 from .enrollment import EnrollmentManager
+from .entity import AccessEntity
 
 
 async def async_setup_entry(